+++ src/Plainion.Prism/Events/ApplicationShutdownEvent.cs
@@ -1,8 +1,0 @@
-﻿using Prism.Events;
-
-namespace Plainion.Prism.Events
-{
-    public class ApplicationShutdownEvent : PubSubEvent<object>
-    {
-    }
-}